make DestructiveConfirmation reusable template
