+++ aries-site/src/examples/templates/DestructiveConfirmation/index.js
@@ -1,0 +1,1 @@
+export * from './DestructiveConfirmation';
