+++ aries-site/src/examples/templates/index.js
@@ -1,4 +1,4 @@
-export * from './ActionMenu';
+/* Template page examples */
 export * from './ascending-navigation';
 export * from './code-blocks';
 export * from './content-layouts';
@@ -14,6 +14,9 @@ export * from './page-layouts';
 export * from './table-customization';
 export * from './toast-notifications';
 export * from './wizard';
+/* Reusable templates */
+export * from './ActionMenu';
+export * from './DestructiveConfirmation';
 export * from './FilterControls';
 export * from './ModalDialog';
 export * from './ReverseAnchor';
