+++ aries-site/src/pages/components/menu.mdx
@@ -184,7 +184,7 @@ Actions enabling a user to share or consume data related to a resource.
 
 #### Batch actions menu related to a selection of servers
 
-The table menu's batch actions are grouped by exectuing and configuration actions. 
+The table menu's batch actions are grouped by executing and configuration actions. 
 The executing actions group is ordered first because the actions in it are the 
 most frequently used.
